feat: fetch 역할을 하는 Client 추가 및 적용
