+++ src/index.ts
@@ -1,13 +1,10 @@
 import './assets/common.css';
+import client from './client';
 import { MoviesLoader } from './MoviesLoader';
 
-const getPopularMoviesURL = (page: number) =>
-  `https://api.themoviedb.org/3/movie/popular?api_key=${process.env.TMDB_API_KEY}&language=ko-KR&page=${page}`;
+const fetchFn = (page: number) => client.getPopularMovies(page);
 
-const getSearchMoviesURL = (query: string) => (page: number) =>
-  `https://api.themoviedb.org/3/search/movie?api_key=${process.env.TMDB_API_KEY}&language=ko-KR&query=${query}&page=${page}&include_adult=false`;
-
-let moviesLoader = new MoviesLoader(getPopularMoviesURL);
+let moviesLoader = new MoviesLoader(fetchFn);
 
 document.querySelector('.search-box')?.addEventListener('submit', (event) => {
   event.preventDefault();
@@ -15,7 +12,7 @@ document.querySelector('.search-box')?.addEventListener('submit', (event) => {
   const formData = Object.fromEntries(new FormData(event.target as HTMLFormElement).entries());
   const query = formData['search-text'] as string;
 
-  moviesLoader = new MoviesLoader(getSearchMoviesURL(query));
+  moviesLoader = new MoviesLoader((page: number) => client.getSearchMovies(query, page));
 });
 
 document.querySelector('#item-load')?.addEventListener('click', (event) => {
