+++ src/movies.type.ts
@@ -1,0 +1,5 @@
+export type Movie = {
+  title: string;
+  vote_average: number;
+  poster_path: string;
+};
